refactor-#156: move customer query to queries
