+++ src/microservices/customer/graphql/queries/index.ts
@@ -1,0 +1,1 @@
+export * from './customer-query.resolver';
